LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-channelmergernode-interface/audiochannelmerger-input-expected.txt

CONSOLE MESSAGE: line 18: TypeError: undefined is not an object (evaluating 'context.startRendering().then')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'context.startRendering().then')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [silent-channel]  
